# openapi/heroscript-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HeroScript API",
    "version": "v1",
    "description": "Contrato público para runs determinísticas, conteúdo imutável e replay verificável. Toda resposta não bem-sucedida usa RFC 9457 Problem Details."
  },
  "servers": [
    { "url": "http://localhost:5260", "description": "Desenvolvimento local" }
  ],
...
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "tags": ["Runs"],
        "operationId": "listRuns",
        "x-heroscript-stability": "stable",
        "responses": { "200": { "description": "Persisted run summaries" } }
      },
      "post": {
        "tags": ["Runs"],
        "operationId": "startRun",
        "x-heroscript-stability": "stable",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RunStartRequest" } } }
        },
        "responses": {
          "200": { "description": "Run created deterministically" },
          "400": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
    "/api/v1/combats/{combatId}": {
      "get": {
        "tags": ["Combats"],
        "operationId": "getCombat",
        "x-heroscript-stability": "stable",
        "parameters": [{ "$ref": "#/components/parameters/CombatId" }],
        "responses": { "200": { "description": "Combat read model" }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/api/v1/combats/{combatId}/history": {
      "get": {
        "tags": ["Combats"],
        "operationId": "getCombatHistory",
        "x-heroscript-stability": "stable",
        "parameters": [{ "$ref": "#/components/parameters/CombatId" }],
        "responses": { "200": { "description": "Resolved combat action history" } }
      }
    },
...
    "/api/v1/sandbox/scenarios/validate": {
      "post": {
        "tags": ["Sandbox"],
        "operationId": "validateCombatScenario",
        "x-heroscript-stability": "experimental",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CombatScenarioDefinition" } } } },
        "responses": { "200": { "description": "Normalized, validated scenario and deterministic scenario hash" }, "422": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/v1/sandbox/runs": {
      "post": {
        "tags": ["Sandbox"],
        "operationId": "launchCombatScenario",
        "x-heroscript-stability": "experimental",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CombatScenarioDefinition" } } } },
        "responses": { "200": { "description": "Idempotent scenario run and active combat launch" }, "422": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/v1/sandbox/runs/{runId}/scenario": {
      "get": {
        "tags": ["Sandbox"],
        "operationId": "getCombatScenario",
        "x-heroscript-stability": "experimental",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "Persisted normalized scenario input" }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/v1/sandbox/runs/{runId}/snapshot": {
      "get": {
        "tags": ["Sandbox"],
        "operationId": "getCombatSandboxSnapshot",
        "x-heroscript-stability": "experimental",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "Consolidated read-only run, combat, actor and card-instance projection" }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    }
  },
  "components": {
    "parameters": {
      "RunId": { "name": "runId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "CombatId": { "name": "combatId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem response",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      }
    },
    "securitySchemes": {
      "AdminApiKey": { "type": "apiKey", "in": "header", "name": "X-Admin-Key" }
    },
    "schemas": {
      "RunStartRequest": {
        "type": "object",
        "required": ["configName", "runDefinitionId", "playerEntityId"],
        "properties": {
          "configName": { "type": "string" },
          "runDefinitionId": { "type": "string" },
          "playerEntityId": { "type": "string" },
          "seed": { "type": "integer", "format": "uint64" },
          "contentRevision": { "type": "string" },
          "modeId": { "type": "string", "default": "standard", "description": "Published game mode; omitted selects the canonical standard mode." }
        }
      },
      "CombatScenarioDefinition": {
        "type": "object",
        "required": ["schemaVersion", "modeId", "seed", "attemptKey", "participants", "deck"],
        "properties": {
          "schemaVersion": { "type": "integer", "const": 2 },
          "modeId": { "type": "string" },
          "contentRevision": { "type": "string", "description": "Published content revision; omitted means the current revision." },
          "seed": { "type": "integer", "format": "uint64" },
          "attemptKey": { "type": "string", "description": "Changes the deterministic identity of a new attempt." },
          "participants": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["instanceId", "entityDefinitionId", "sideId", "controllerBinding"],
              "properties": {
                "instanceId": { "type": "string" },
                "entityDefinitionId": { "type": "string" },
                "sideId": { "type": "string" },
                "controllerBinding": {
                  "type": "object",
                  "required": ["kind"],
                  "properties": {
                    "kind": { "type": "string", "enum": ["Player", "AI", "None"] },
                    "policyId": { "type": ["string", "null"] }
                  },
                  "additionalProperties": false
                }
              },
              "additionalProperties": false
            }
          },
          "deck": { "type": "array", "items": { "type": "object" } },
          "initialState": { "type": "object" }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "code", "correlationId"],
        "properties": {
          "type": { "type": "string", "format": "uri-reference" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri-reference" },
          "code": { "type": "string", "description": "Stable machine-readable HeroScript error code" },
          "correlationId": { "type": "string", "description": "Request correlation identifier" },
          "currentSequence": { "type": "integer", "description": "Authoritative sequence when a command conflicts" },
          "currentStep": { "type": "integer", "format": "uint64", "description": "Authoritative deterministic step when a command conflicts" }
        }
      }
    }
  }
}